Create comparing_the_triplets.py
In this solution, I define a function called compareTriplets that takes in two lists a and b as arguments. I initialize alice and bob to 0, and then loop through the lists using a for loop.

For each index, I compare the corresponding elements of a and b. If a[i] is greater than b[i], it add 1 to alice. If b[i] is greater than a[i], it add 1 to bob. If they are equal, we don't add any points.

Finally it return the result as a list [alice, bob] points