A Very Big Sum
The aVeryBigSum function takes a list of integers ar as input and calculates the sum of all the elements in the list. It does this by initializing a variable named result to zero and then iterating through the list using a for loop. During each iteration, the current element is added to the result variable using the += operator. Finally, the function returns the result variable, which contains the sum of all the elements in the input list.